+++ selcukes-databind/src/test/java/io/github/selcukes/databind/tests/JsonUtilsTest.java
@@ -1,0 +1,38 @@
+/*
+ *  Copyright (c) Ramesh Babu Prudhvi.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+package io.github.selcukes.databind.tests;
+
+import io.github.selcukes.databind.utils.JsonUtils;
+import org.testng.annotations.Test;
+
+import java.util.LinkedHashMap;
+
+import static org.testng.Assert.assertEquals;
+
+public class JsonUtilsTest {
+
+    @Test
+    public void jsonTest() {
+        var map = new LinkedHashMap<>();
+        map.put("a", "1");
+        map.put("b", "2");
+        map.put("c", "3");
+        String expected = "{\"a\":\"1\",\"b\":\"2\",\"c\":\"3\"}";
+        assertEquals(JsonUtils.toJson(map), expected);
+    }
+
+}
